repos/distributedstatemachine/gcloud_hub/server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "server code for hubezkl-backend"
authors = ["jseam2 <hello.jseam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
strawberry-graphql = "^0.201.1"
flask = "^2.3.2"
rich = "^13.4.2"
typer = "^0.9.0"
libcst = "^1.0.1"
flask-cors = "^4.0.0"
flask-sqlalchemy = "^3.0.5"
psycopg2-binary = "^2.9.6"
flask-migrate = "^4.0.4"
flask-wtf = "^1.1.1"
celery = {extras = ["redis"], version = "^5.3.1"}
gunicorn = "^21.2.0"
nest-asyncio = "^1.5.7"
authlib = "^1.2.1"
requests = "^2.31.0"
cryptography = "^41.0.3"
pytest = "^7.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ruff = "^0.0.278"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
